+++ Sora/PeerChannel.swift
@@ -1,12 +1,6 @@
 import Foundation
 import WebRTC
 
-/**
- ピアチャネルのイベントハンドラです。
- */
-@available(*, unavailable, message: "MediaChannelHandlers を利用してください。")
-public class PeerChannelHandlers {}
-
 final class PeerChannelInternalHandlers {
     /// 接続解除時に呼ばれるクロージャー
     var onDisconnect: ((Error?, DisconnectReason) -> Void)?
@@ -113,7 +107,7 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
     var bundleId: String?
     var connectionId: String?
 
-    var onConnectHandler: ((Error?) -> Void)?
+    var onConnect: ((Error?) -> Void)?
 
     var isAudioInputInitialized: Bool = false
 
@@ -171,7 +165,7 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
 
         // このロックは finishConnecting() で解除される
         lock.lock()
-        onConnectHandler = handler
+        onConnect = handler
 
         // サイマルキャストを利用する場合は、 RTCPeerConnection の生成前に WrapperVideoEncoderFactory を設定する必要がある
         // また、スポットライトはサイマルキャストを利用しているため、同様に設定が必要になる
@@ -233,8 +227,8 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
         if let error {
             Logger.error(type: .peerChannel,
                          message: "failed connecting to signaling channel (\(error.localizedDescription))")
-            onConnectHandler?(error)
-            onConnectHandler = nil
+            onConnect?(error)
+            onConnect = nil
             return
         }
 
@@ -274,15 +268,12 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
         var role: SignalingRole
         var multistream = configuration.multistreamEnabled || configuration.spotlightEnabled == .enabled
         switch configuration.role {
-        case .publisher, .sendonly:
+        case .sendonly:
             role = .sendonly
-        case .subscriber, .recvonly:
+        case .recvonly:
             role = .recvonly
-        case .group, .sendrecv:
+        case .sendrecv:
             role = .sendrecv
-            multistream = true
-        case .groupSub:
-            role = .recvonly
             multistream = true
         }
 
@@ -876,10 +867,10 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
         Logger.debug(type: .peerChannel,
                      message: "native receivers = \(nativeChannel?.receivers.count ?? 0)")
 
-        if onConnectHandler != nil {
+        if onConnect != nil {
             Logger.debug(type: .peerChannel, message: "call connect(handler:)")
-            onConnectHandler!(nil)
-            onConnectHandler = nil
+            onConnect!(nil)
+            onConnect = nil
         }
         lock.unlock()
     }
@@ -909,10 +900,10 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
         Logger.debug(type: .peerChannel, message: "call onDisconnect")
         internalHandlers.onDisconnect?(error, reason)
 
-        if onConnectHandler != nil {
+        if onConnect != nil {
             Logger.debug(type: .peerChannel, message: "call connect(handler:)")
-            onConnectHandler!(error)
-            onConnectHandler = nil
+            onConnect!(error)
+            onConnect = nil
         }
 
         Logger.debug(type: .peerChannel, message: "did disconnect")